tools: perf: split lib.figure.base.Figure
- lib.figure.html - for HTML data tables
- lib.figure.image - for PNG outputs
