+++ tools/perf/CMakeLists.txt
@@ -16,7 +16,6 @@ add_cstyle_pep8(tools_perf
 	${CMAKE_CURRENT_SOURCE_DIR}/lib/bench.py
 	${CMAKE_CURRENT_SOURCE_DIR}/lib/remote_cmd.py
 	${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/__init__.py
-	${CMAKE_CURRENT_SOURCE_DIR}/lib/figure/__init__.py
-	${CMAKE_CURRENT_SOURCE_DIR}/lib/figure/flat.py
+	${CMAKE_CURRENT_SOURCE_DIR}/lib/figure/*.py
 	# ${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/base.py
 	${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/runner/*.py)
