+++ tools/perf/lib/figure/base.py
@@ -8,12 +8,10 @@
 
 import json
 import os.path
-from textwrap import wrap
-from collections import deque
-import matplotlib.pyplot as plt
-from matplotlib.ticker import ScalarFormatter
 from deepdiff import DeepDiff
 
+from .image import draw_png
+from .html import data_table
 from ..common import json_from_file, escape
 
 SKIP_NO_AXIS_MSG = """SKIP: Axis '{}' is not provided by the series of id={}. Available keys are: {}"""
@@ -22,9 +20,6 @@ SKIP_NO_ROWS_MSG = "SKIP: No row is provided by the series of id={}."
 
 class Figure:
     """A single figure"""
-
-    _figure_kwargs = {'figsize': [6.4, 4.8], 'dpi': 200, \
-        'tight_layout': {'pad': 1}}
 
     def __series_file(self, result_dir):
         return os.path.join(result_dir, self.file + '.json')
@@ -77,7 +72,7 @@ class Figure:
         self.yaxis_max = max_y
 
     # a list of possible common params
-    COMMON_PARAMS = {
+    __COMMON_PARAMS = {
         'threads': {
             'default': 1,
             'format': '{}'
@@ -100,14 +95,15 @@ class Figure:
             # use values from the first of the rows. If a given key is missing
             # the default value is used instead.
             params = {key: rows[0].get(key, value['default'])
-                      for key, value in Figure.COMMON_PARAMS.items()}
+                      for key, value in Figure.__COMMON_PARAMS.items()}
         for row in rows:
             # Look row-by-row and create a subset of parameters that are common
             # to all of the rows. If a given key is missing in any of the rows
             # the default value is used instead.
-            params = {key: value for key, value in params.items()
-                      if value == row.get(key,
-                                          Figure.COMMON_PARAMS[key]['default'])}
+            params = {key: value \
+                for key, value in params.items() \
+                if value == row.get(key, \
+                                    Figure.__COMMON_PARAMS[key]['default'])}
         return params
 
     def prepare_series(self, result_dir=None):
@@ -171,153 +167,32 @@ class Figure:
         self.series = output['series']
         self.common_params = output['common_params']
 
-    @staticmethod
-    def _points_to_xy(points):
-        xslist = [p[0] for p in points]
-        yslist = [p[1] for p in points]
-        return xslist, yslist
-
-    def _get_xcommon(self):
+    @property
+    def xcommon(self):
         """generate an ordered list of common x-values"""
         xlist = [p[0] for oneseries in self.series for p in oneseries['points']]
         return sorted(list(set(xlist)))
-
-    @staticmethod
-    def _points_to_dict(points):
-        """transform list of [x, y] into a dict() where {x: y}"""
-        return {p[0]: p[1] for p in points}
-
-    @staticmethod
-    def _label(column, with_better=False):
-        """Translate the name of a column to a label with a unit"""
-        label_by_column = {
-            'threads': '# of threads',
-            'iodepth': 'iodepth',
-            'bs': 'block size [B]',
-            'lat_avg': 'latency [usec]',
-            'lat_pctl_99.9': 'latency [usec]',
-            'lat_pctl_99.99': 'latency [usec]',
-            'bw_avg': 'bandwidth [Gb/s]',
-            'cpuload': 'CPU load [%]'
-        }
-        lower = 'lower is better'
-        higher = 'higher is better'
-        better_by_column = {
-            'lat_avg': lower,
-            'lat_pctl_99.9': lower,
-            'lat_pctl_99.99': lower,
-            'bw_avg': higher
-        }
-        # If the column is not in the dictionary
-        # the default return value is the raw name of the column.
-        output = label_by_column.get(column, column)
-        if with_better:
-            output += '\n(' + better_by_column.get(column, column) + ')'
-        return output
 
     def png_path(self):
         """get a path to the output PNG file"""
         output = self.file + '_' + self.key + '.png'
         return os.path.join('.', output)
 
-    LINE_STYLES = ['solid', 'dashed', 'dashdot', 'dotted']
-
-    @staticmethod
-    def draw_png(argx, argy, series, xscale, output_path, yaxis_max=None,
-                 suptitle=None, title=None, oneseries_name='label'):
-        """draw a figure"""
-        # set output file size, padding and title
-        fig = plt.figure(**Figure._figure_kwargs)
-        if suptitle is not None:
-            suptitle = "\n".join(wrap(suptitle, 60))
-            fig.suptitle(suptitle, fontsize='medium', y=0.90)
-        # get a subplot
-        plot = plt.subplot(1, 1, 1)
-        if title is not None:
-            plot.title.set_text(title)
-            plot.title.set_fontsize(10)
-        xticks = []
-        line_styles = deque(Figure.LINE_STYLES.copy())
-        group_to_line_styles = {}
-        for oneseries in series:
-            # Pick a line style according to the group to which
-            # the line belongs. If no group provided a default one is assumed.
-            group = oneseries.get('group', 'default')
-            if group in group_to_line_styles:
-                line_style = group_to_line_styles[group]
-            else:
-                if len(line_styles) > 0:
-                    line_style = line_styles.popleft()
-                else:
-                    raise Exception('Too many benchmarks.')
-                group_to_line_styles[group] = line_style
-            # draw series ony-by-one
-            xslist, yslist = Figure._points_to_xy(oneseries['points'])
-            plot.plot(xslist, yslist, marker='.', linestyle=line_style,
-                      label=oneseries[oneseries_name])
-            # collect all existing x values
-            xticks.extend(xslist)
-        # make values unique (set) and sort them
-        xticks = sorted(list(set(xticks)))
-        # set the x-axis scale
-        if xscale == "linear":
-            plot.set_xscale(xscale)
-        else:
-            plot.set_xscale(xscale, base=2)
-            plot.xaxis.set_major_formatter(ScalarFormatter())
-
-        plot.set_xticks(xticks)
-        plt.setp(plot.get_xticklabels(), rotation=45, ha='right')
-        plot.set_xlabel(Figure._label(argx))
-        plot.set_ylabel(Figure._label(argy, with_better=True))
-        if yaxis_max is not None:
-            plot.set_ylim(top=yaxis_max)
-        plot.set_ylim(bottom=0)
-        plot.legend(fontsize=10)
-        plot.grid(True)
-
-        plt.savefig(output_path)
-        plt.close(fig)
-
     def to_png(self, include_title):
         """generate an output PNG file"""
         os.chdir(self.result_dir)
         suptitle = self.title if include_title else None
         title = '[{}]'.format(', '.join(['{}={}'. \
-            format(key, Figure.COMMON_PARAMS[key]['format'].format(value))
+            format(key, Figure.__COMMON_PARAMS[key]['format'].format(value))
                                          for key, value in
                                          self.common_params.items()]))
-        Figure.draw_png(self.argx, self.argy, self.series, self.xscale,
-                        self.png_path(), self.yaxis_max, suptitle, title)
-
-    def html_data_table(self):
-        """
-        Create an HTML snippet string with a table containing the Figure data.
-        """
-        # header
-        xcommon = self._get_xcommon()
-        html = '<table class="data"><thead><tr><th></th><th>{}</th></tr>' \
-            '</thead><tbody>'. \
-            format('</th><th>'.join([str(x) for x in xcommon]))
-        # rows
-        for oneseries in self.series:
-            # Since the output is processed as markdown,
-            # special characters have to be escaped.
-            html += "<tr><td>" + escape(oneseries['label']) + "</td>"
-            points = Figure._points_to_dict(oneseries['points'])
-            points = {k: '{0:.2f}'.format(v) for k, v in points.items()}
-            for xarg in xcommon:
-                html += '<td>{}</td>'.format(points.get(xarg, '-'))
-            html += "</tr>"
-
-        # end the table
-        html += "</tbody></table>"
-        return html
+        draw_png(self.argx, self.argy, self.series, self.xscale,
+                 self.png_path(), self.yaxis_max, suptitle, title)
 
     def to_html(self, figno):
         """Combine a Figure's png and data table into a single HTML snippet"""
         html = "<h4 class='figure'>Figure {}. {}</h4>". \
             format(figno, escape(self.title))
         html += '<img src="' + self.png_path() + '" alt="' + self.title + '"/>'
-        html += self.html_data_table()
+        html += data_table(self.xcommon, self.series)
         return html
